PageParser: Get rid of the PageParserState struct entirely and use docs instead
The purpose of PageParserState was to track proper order of function calls.

However, this still was not 100% safe as e.g. when one function failed, it may
have edited the internal state, but not invalidated the state, thus allowing
for re-runs of that function, returning invalid data.

Instead we document the intended order of calling, and hope that users
abide it.