Fix seeking to before the final page once it has been read
Re-implements the finding of the initial boundaries via
a "clean slate" approach where each seek starts equally
regardless of the position in the current stream.

This is of course disadvantageous in some cases, e.g.
when we are close to a border and want to seek somewhere
between our current position and the border we are close
to.