Restore compatibility with older versions of Rust
Newer versions of Rust allow omitting the field name if it is equal to
the variable name. This feature was stabilized in Rust 1.17. By
including the field name again, this restores compatibility with Rust
versions at least as old as 1.13 (I did not try older versions). For
such a small change, the compatibility win seems worth the extra bytes.