Add repack example
E.g. this can be used to fix CRC checksums in files obtained
by fuzzing methods.
So if you do this:

RUSTFLAGS="--cfg fuzzing" cargo run --example repack <A> <B>

it will read from A and repack it to B while ignoring any CRC
errors.

Note that the repack example is atm not expected to reproduce
100% the same file: page boundaries are not being respected
so they might turn out differently in the generated file.
This might then in turn mess up the packet <-> absgp link.

But to fix all of this we'd have to add a "end of page" flag
to packets.