Reorganize Packet struct interface and add in-page info
Firstly, we add info about whether a given packet is
the last one ending inside a page or the first one
starting.

Secondly, we redesign the packet struct contents and
make most of them private, only accessible through
functions. This serves the purpose to make such
additions doable in the future in a backwards
compatible fashion, as adding fields to a struct
full of public only fields is not backwards
compatible.
This smart idea is also recommended by the rust API
guidelines [1].

[1]: https://rust-lang-nursery.github.io/api-guidelines/future-proofing.html#structs-have-private-fields-c-struct-private