Remove deny(warnings) and add github actions CI
Back in the day I haven't set up CI because I didn't think
it important for this small crate. Now my views on it have
changed, I think CI is a good idea.
How you change as a person :).

We now deny warnings in CI if the toolchain matches the
MSRV, and remove deny(warnings) from CI. It was warning
about various deprecations on newer rustcs. This should
make it possible for crater to detect the ogg crate as
test success. IIRC it's also suggested that people do
this.