Let PacketWriter::write_packet use buffers convertible to Cow
As pointed out in issue #19, in some cases it may not be desirable to
heap allocate Ogg packet contents, but the current parameters of the
`write_packet` method impose a heap allocation no matter what.

However, the `PacketWriter` struct doesn't actually need to own the
packet data, as long as it stays in memory unmodified during its
lifetime. The `Cow` smart pointer that the standard library provides
captures this requirement more precisely, and provides for more elegant,
flexible and potentially efficient client code.

Therefore, substitute `Box` for `Cow` in the aforementioned code. Boxed
slices can still be used with this new design, although with a different
syntax, which breaks backwards compatibility.