Add the async API to the documentation generated by docs.rs
Currently, the documentation available at docs.rs for this crate does
not mention the feature-gated async API at all, which makes it harder
for users to discover it. I think everyone can agree that documenting it
better is a good thing.

However, documenting feature-gated APIs is not that straightforward, as
the naive approach of passing `--all-features` to `cargo doc` generates
documentation that does not mention under which conditions feature-gated
APIs are available, and resorting to unstable rustdoc features means
that the project can no longer be developed on a stable Rust toolchain.

As a compromise but fairly elegant solution, exploit the fact that the
documentation most end-users will consume is generated by docs.rs, which
uses a nightly toolchain and allows setting conditional compilation
options to only enable unstable rustdoc features in that environment.

People interested in building their own documentation locally that would
look exactly as the docs.rs one can do so by using a nightly toolchain
and replicating the docs.rs environment configuration. The following
command does that:

```
RUSTDOCFLAGS="--cfg docsrs" cargo +nightly doc --all-features
```

While at it, also change the documentation link in the README to use a
prettier badge.