Allow instantiating async `PacketReader` with a futures `AsyncRead` (#31)
* Allow instantiating async PacketReader with a futures AsyncRead

The async `PacketReader` struct is implemented using `tokio`, which
nowadays is the _de facto_ standard async runtime in the Rust ecosystem.
However, other async runtimes exist, such as `async-std` and `smol`, and
people may want to use these alternatives for a variety of reasons.

This internal dependency on Tokio is noticeable to end-users of the async
`PacketReader` struct: they need to pass a Tokio `AsyncRead` no matter
what, even though each async runtime has its own similar, but not quite
the same implementation of the `AsyncRead` trait.

The futures crate is said to contain types meant to be interoperable and
maybe eventually integrated in the Rust standard library, so accepting
its implementation of `AsyncRead` looks like the most compatible choice
across the whole async ecosystem. It also is easier for end-users to
wrap whatever incompatible `AsyncRead` they might be using to a futures
`AsyncRead`, as depending on the futures crates is more lightweight and
feels more like "neutral ground".

Therefore, add a new compatibility constructor for `PacketReader` that
accepts a futures `AsyncRead`, and then uses the Tokio compatibility
layer to get a Tokio `AsyncRead` from it. I think it's best to not drop
the previous Tokio `AsyncRead` constructor for two reasons:

- As the Tokio runtime is the _de facto_ standard, it is reasonable to
  care about the ergonomics of using its `AsyncRead`s. Tokio offers a
  compatibility layer to go from its `AsyncRead` to the futures
  `AsyncRead`, but asking client code to use it is more cumbersome and a
  breaking API change.
- The compatibility layer may have a small performance penalty, due to
  Tokio no longer being able to safely assume some invariants that its
  implementations of the `AsyncRead` trait hold.

* Fix build failure when running async integration tests

These really need the futures_util dependency that was removed.

* Add empty white line in rustdoc comments as per review