Add async PacketWriter
Given that this crate already supports async reads, it makes sense to
complete its async support for writes for API parity.

To achieve this addition the packet writing logic was extracted to a new
private struct, `BasePacketWriter`, that is largely the same as before,
with the exception of some refactors to delete data from ended streams
of the map, which was a TODO comment, and getting rid of byteorder and
fallible I/O operations. This new struct is not responsible for doing
any I/O: it just invokes a callback, which can return `false` to
indicate a failure and stop the writing process.

The blocking `PacketWriter` now uses `BasePacketWriter` with a callback
that writes everything to the sink as before. Any error that might
happen is stored and returned, with the same user-visible behavior as
before.

The new async `PacketWriter` is pretty trivial thanks to Tokio encoders
and the usage of callbacks by `BasePacketWriter`, as it is possible to
simply copy any written pages to a Tokio-provided buffer, which always
succeeds. Tokio takes care of all the dirty details of actual I/O.