Do not return error on garbage data after last Ogg page (#34)
* Do not return error on garbage data after last Ogg page

Some admittely badly-encoded Ogg Vorbis files in the wild contain
trailing garbage data after the last Ogg page, such as this one:
https://freesound.org/people/stomachache/sounds/157587/.

The page reading code explicitly handled this case by returning an
error, which makes sense to do given that the Ogg specification is not
clear on how this condition should be dealt with. Moreover, programs
like `ogginfo` also show warnings for these kind of files. However,
being this strict causes codec crates that depend on this one to not be
able to play back such files properly, as discussed in
https://github.com/RustAudio/ogg/issues/7.

Fix the situation by being more lenient with this nonsense. Some
additional status information is kept to distinguish this case from
empty files.

* Add unit test for proper trailing data after last page handling

* Ensure that trailing data is not returned in a packet in unit test

I don't expect this additional check to be useful in practice, but it is
an invariant that should be held in the test case after all.