Add method to retrieve the checksum of the end page for a `Packet`
The Ogg specification demands that each Ogg page has a checksum of its
data for error detection and correction purposes. Logically, this
checksum is an implementation detail of the Ogg transport, and
higher-level decoders need not know it to parse logical bitstreams.
However, specialized decoders may want to efficiently use it to compute
a checksum of the data that flows through it, avoiding hashing packet
data again. In addition, testing code may also find use in a checksum to
asser that the data of the page a packet belongs to is reasonably
similar to what's expected.

Given that the Ogg parsing code already reads and compares page
checksums, exposing them in the public `Packet` API, which already
allows users to retrieve Ogg encapsulation data such as the granule
position, comes at practically no additional cost.